Projects/CyberGuard/requirements.txt
tensorflow==2.9.1
numpy==1.23.1
pandas==1.4.3
scikit-learn==1.1.1
flask==2.1.3
flask-cors==3.0.10
pymongo==4.2.0
requests==2.28.1
python-dotenv==0.20.0
pyyaml==6.0
pytest==7.1.2
